Refactor into features submodule
