+++ snorkel/features/__init__.py
@@ -1,0 +1,8 @@
+"""
+Subpackage for Snorkel featurization.
+"""
+
+from .relative_features import *
+from .context_features import *
+from .entity_features import *
+from .generic_features import feats_from_matrix_generator
